Tidy up Bones code
