+++ ScriptHookVDotNet.vcxproj.filters
@@ -248,9 +248,6 @@
     <CsCompile Include="source\scripting\DlcWeaponStructs.cs">
       <Filter>scripting\World\Weapons</Filter>
     </CsCompile>
-    <CsCompile Include="source\scripting\Bone.cs">
-      <Filter>scripting\World\Entities\Peds</Filter>
-    </CsCompile>
     <CsCompile Include="source\scripting\Tasks.cs">
       <Filter>scripting\World\Entities\Peds</Filter>
     </CsCompile>
@@ -262,6 +259,18 @@
     </CsCompile>
     <CsCompile Include="source\scripting\Style.cs">
       <Filter>scripting\World\Entities\Peds</Filter>
+    </CsCompile>
+    <CsCompile Include="source\scripting\BoneID.cs">
+      <Filter>scripting\World\Entities\Peds</Filter>
+    </CsCompile>
+    <CsCompile Include="source\scripting\PedBone.cs">
+      <Filter>scripting\World\Entities\Peds</Filter>
+    </CsCompile>
+    <CsCompile Include="source\scripting\PedBoneCollection.cs">
+      <Filter>scripting\World\Entities\Peds</Filter>
+    </CsCompile>
+    <CsCompile Include="source\scripting\EntityBoneCollection.cs">
+      <Filter>scripting\World\Entities</Filter>
     </CsCompile>
   </ItemGroup>
 </Project>
