+++ source/scripting/EntityBone.cs
@@ -1,10 +1,4 @@
 ﻿using System;
-using System.CodeDom;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using GTA.Math;
 using GTA.Native;
 
@@ -298,262 +292,4 @@ namespace GTA
 			return new InputArgument(entityBone.Index);
 		}
 	}
-
-	public class PedBone : EntityBone
-	{
-		#region Fields
-		private new readonly Ped _owner;
-		#endregion
-
-		public new Ped Owner
-		{
-			get { return _owner;}
-		}
-
-		internal PedBone(Ped owner, int boneIndex) : base(owner, boneIndex)
-		{
-			_owner = owner;
-		}
-
-		internal PedBone(Ped owner, string boneName) : base(owner, boneName)
-		{
-			_owner = owner;
-		}
-
-		internal PedBone(Ped owner, Bone boneId)
-			: base(owner, Function.Call<int>(Hash.GET_PED_BONE_INDEX, owner.Handle, boneId))
-		{
-			_owner = owner;
-		}
-
-	    public new bool IsValid
-	    {
-	        get
-	        {
-	            return Ped.Exists(Owner) && Index != -1;
-	        }
-	    }
-
-	}
-
-	public class EntityBoneCollection : IEnumerable<EntityBone>
-	{
-        public class Enumerator : IEnumerator<EntityBone>
-        {
-            #region Fields
-            private readonly EntityBoneCollection _collection;
-            private int currentIndex = -1;// skip the CORE bone index(-1)
-            #endregion
-            public Enumerator(EntityBoneCollection collection)
-            {
-                _collection = collection;
-            }
-            public EntityBone Current
-            {
-                get { return _collection[currentIndex]; }
-            }
-
-            object IEnumerator.Current
-            {
-                get { return _collection[currentIndex]; }
-            }
-
-            public void Dispose()
-            {
-
-            }
-
-            public bool MoveNext()
-            {
-                return ++currentIndex < _collection.Count;
-            }
-
-            public void Reset()
-            {
-                currentIndex = -1;
-            }
-        }
-
-        #region Fields
-        protected readonly Entity _owner;
-        #endregion
-
-        internal EntityBoneCollection(Entity owner)
-		{
-			_owner = owner;
-		}
-
-		/// <summary>
-		/// Gets the <see cref="EntityBone"/> with the specified bone name.
-		/// </summary>
-		/// <param name="boneName">Name of the bone.</param>
-		public EntityBone this[string boneName]
-		{
-			get { return new EntityBone(_owner, boneName); }
-		}
-
-		/// <summary>
-		/// Gets the <see cref="EntityBone"/> at the specified bone index.
-		/// </summary>
-		/// <param name="boneIndex">The bone index.</param>
-		public EntityBone this[int boneIndex]
-		{
-			get { return new EntityBone(_owner, boneIndex); }
-		}
-
-		/// <summary>
-		/// Determines whether this <see cref="Entity"/> has a bone with the specified bone name
-		/// </summary>
-		/// <param name="boneName">Name of the bone.</param>
-		/// <returns>
-		///   <c>true</c> if this <see cref="Entity"/> has a bone with the specified bone name; otherwise, <c>false</c>.
-		/// </returns>
-		public bool HasBone(string boneName)
-		{
-			return Function.Call<int>(Hash.GET_ENTITY_BONE_INDEX_BY_NAME, _owner.Handle, boneName) !=  -1;
-		}
-
-		/// <summary>
-		/// Gets the number of bones that this <see cref="Entity"/> has.
-		/// </summary>
-		public int Count
-		{
-			get { return MemoryAccess.GetEntityBoneCount(_owner.Handle); }
-		}
-
-		public EntityBone Core
-		{
-			get { return new EntityBone(_owner, -1); }
-		}
-
-	    public IEnumerator<EntityBone> GetEnumerator()
-	    {
-	        return new Enumerator(this);
-	    }
-
-	    public override int GetHashCode()
-		{
-			return _owner.GetHashCode() ^ Count.GetHashCode();
-		}
-
-	    IEnumerator IEnumerable.GetEnumerator()
-	    {
-	        return GetEnumerator();
-	    }
-	}
-
-	public class PedBoneCollection : EntityBoneCollection, IEnumerable<PedBone>
-	{
-        public new class Enumerator : IEnumerator<PedBone>
-        {
-            #region Fields
-            private readonly PedBoneCollection _collection;
-            private int currentIndex = -1;// skip the CORE bone index(-1)
-            #endregion
-            public Enumerator(PedBoneCollection collection)
-            {
-                _collection = collection;
-            }
-            public PedBone Current
-            {
-                get { return _collection[currentIndex]; }
-            }
-
-            object IEnumerator.Current
-            {
-                get { return _collection[currentIndex]; }
-            }
-
-            public void Dispose()
-            {
-
-            }
-
-            public bool MoveNext()
-            {
-                return ++currentIndex < _collection.Count;
-            }
-
-            public void Reset()
-            {
-                currentIndex = -1;
-            }
-        }
-
-        #region Fields
-        private new readonly Ped _owner;
-		#endregion
-		internal PedBoneCollection(Ped owner) : base(owner)
-		{
-			_owner = owner;
-		}
-
-		/// <summary>
-		/// Gets the <see cref="PedBone"/> with the specified bone name.
-		/// </summary>
-		/// <param name="boneName">Name of the bone.</param>
-		public new PedBone this[string boneName]
-		{
-			get { return new PedBone(_owner, boneName); }
-		}
-
-		/// <summary>
-		/// Gets the <see cref="PedBone"/> at the specified bone index.
-		/// </summary>
-		/// <param name="boneIndex">The bone index.</param>
-		public new PedBone this[int boneIndex]
-		{
-			get { return new PedBone(_owner, boneIndex); }
-		}
-
-        /// <summary>
-        /// Gets the <see cref="PedBone"/> with the specified boneId.
-        /// </summary>
-        /// <param name="boneId">The boneId.</param>
-        public PedBone this[Bone boneId]
-		{
-			get { return new PedBone(_owner, boneId); }
-		}
-
-	    public new PedBone Core
-	    {
-	        get { return new PedBone(_owner, -1); }
-	    }
-
-		/// <summary>
-		/// Gets the last damaged Bone for this <see cref="Ped"/>.
-		/// </summary>
-		public PedBone LastDamaged
-		{
-			get
-			{
-			    int outBone;
-				unsafe
-				{
-					if (Function.Call<bool>(Hash.GET_PED_LAST_DAMAGE_BONE, _owner.Handle, &outBone))
-					{
-						return this[(Bone) outBone];
-					}
-				}
-				return this[Bone.SKEL_ROOT];
-			}
-		}
-
-		/// <summary>
-		/// Clears the last damage a bone on this <see cref="Ped"/> received.
-		/// </summary>
-		public void ClearLastDamaged()
-		{
-			Function.Call(Hash.CLEAR_PED_LAST_DAMAGE_BONE, _owner.Handle);
-		}
-
-	    public new IEnumerator<PedBone> GetEnumerator()
-	    {
-	        return new Enumerator(this);
-	    }
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
-    }
 }
